requirements.txt
flask>=3.0.0
requests>=2.31.0
anthropic>=0.40.0
pdfplumber>=0.11.0
python-docx>=1.1.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
APScheduler>=3.10.0
pystray>=0.19.0
Pillow>=10.0.0
